msm: ipa: Remove SUSPEND_PREPARE event handling
SUSPEND_PREPARE event handling is not needed as AP gets the
pm_ops callback once the device enters suspend mode. Deactivating
the clients as part of this event is resulting in race conditions.

Revert the mutex check while dea-activating the clients. All the
clients will be in deactivated state while handling this callback
and no state changes are expected.

Change-Id: I67806c89fa18b2c23c1c6f87fb203322361569f5
Signed-off-by: Pavan Kumar M <rpavan@qti.qualcomm.com>
Signed-off-by: GuidixX <guidixm@gmail.com>
Signed-off-by: Guidix <guidixm@gmail.com>